change how deprecation is implemented
it only makes sense to deprecate properties coming from some external
source.  it doesn't make sense to deprecate a value transformer, for
example, because that's operating on a value retrieved from elswhere.
Deprecation should be done at the access to the external source (where
the key and source name are available to provide a better log/exception
anyway)